+++ server/database/index.ts
@@ -1,4 +1,5 @@
 export * from "./auth";
+export * from "./bookmark";
 export * from "./business";
 export * from "./business-hours";
 export * from "./email";